refactor(shared): move useHelpPopover into src/components/shared
Extracts the generic popover-state hook from SettingsOverlay into
src/components/shared/useHelpPopover.ts using a generic TId type param
(Map-based container registry). The SettingsOverlay local file is
replaced with a single re-export shim so all existing imports remain
stable.
